After moving, test works
