+++ django_geosource/tests/test_commands.py
@@ -8,6 +8,10 @@ from rest_framework.exceptions import MethodNotAllowed
 from django_geosource.models import GeoJSONSource, GeometryTypes
 
 
+def side_effect(method, list, **kwargs):
+    return "Task"
+
+
 class ResyncAllSourcesTestCase(TestCase):
     def setUp(self):
         self.source = GeoJSONSource.objects.create(
@@ -17,17 +21,15 @@ class ResyncAllSourcesTestCase(TestCase):
         )
 
     def test_resync_all_sources(self):
-        def side_effect(method, list, **kwargs):
-            return "Task"
 
         with mock.patch(
-            "django_geosource.celery.app.send_task", side_effect=side_effect
-        ) as mocked:
-            with mock.patch(
-                "django_geosource.mixins.CeleryCallMethodsMixin.update_status",
-                return_value=False,
-            ):
-                call_command("resync_all_sources")
+            "django_geosource.models.GeoJSONSource.refresh_data"
+        ) as mocked, mock.patch(
+            "django_geosource.mixins.CeleryCallMethodsMixin.update_status",
+            return_value=False,
+        ):
+            call_command("resync_all_sources")
+
         mocked.assert_called_once()
 
     def test_resync_all_sources_sync(self):
@@ -42,13 +44,12 @@ class ResyncAllSourcesTestCase(TestCase):
             return "Task"
 
         with mock.patch(
-            "django_geosource.celery.app.send_task", side_effect=side_effect
-        ) as mocked:
-            with mock.patch(
-                "django_geosource.mixins.CeleryCallMethodsMixin.update_status",
-                return_value=False,
-            ):
-                call_command("resync_source", pk=self.source.id)
+            "django_geosource.models.GeoJSONSource.refresh_data"
+        ) as mocked, mock.patch(
+            "django_geosource.mixins.CeleryCallMethodsMixin.update_status",
+            return_value=False,
+        ):
+            call_command("resync_source", pk=self.source.id)
         mocked.assert_called_once()
 
     def test_resync_source_sync(self):
@@ -72,19 +73,17 @@ class ResyncAllSourcesTestCase(TestCase):
                     call_command("resync_all_sources")
 
     def test_resync_all_sources_fail_force(self):
-        def side_effect(method, list, **kwargs):
-            return "Task"
 
         with mock.patch(
-            "django_geosource.celery.app.send_task", side_effect=side_effect
-        ) as mocked:
+            "django_geosource.models.GeoJSONSource.refresh_data"
+        ) as mocked, mock.patch(
+            "django_geosource.mixins.CeleryCallMethodsMixin.update_status",
+            return_value=False,
+        ):
             with mock.patch(
-                "django_geosource.mixins.CeleryCallMethodsMixin.update_status"
+                "django_geosource.mixins.CeleryCallMethodsMixin.can_sync",
+                new_callable=mock.PropertyMock,
+                return_value=False,
             ):
-                with mock.patch(
-                    "django_geosource.mixins.CeleryCallMethodsMixin.can_sync",
-                    new_callable=mock.PropertyMock,
-                    return_value=False,
-                ):
-                    call_command("resync_all_sources", force=True)
+                call_command("resync_all_sources", force=True)
         mocked.assert_called_once()
