+++ django_geosource/tests/test_tasks.py
@@ -5,9 +5,7 @@ from unittest import mock
 from django.contrib.auth.models import Group
 from django.test import TestCase
 
-from celery.exceptions import Ignore
-
-from django_geosource.celery.tasks import run_model_object_method
+from django_geosource.tasks import run_model_object_method
 from django_geosource.models import GeoJSONSource, GeometryTypes
 
 from geostore.models import Feature, Layer
@@ -24,13 +22,14 @@ class TaskTestCase(TestCase):
         )
 
     def test_task_refresh_data_method(self):
-        with self.assertRaises(Ignore):
-            run_model_object_method(
+        run_model_object_method.apply(
+            (
                 self.element._meta.app_label,
                 self.element._meta.model_name,
                 self.element.pk,
                 "refresh_data",
             )
+        )
         self.assertEqual(Layer.objects.count(), 1)
         self.assertEqual(Feature.objects.count(), 1)
         self.assertEqual(Feature.objects.first().properties, {"id": 1, "test": 5})
@@ -38,35 +37,38 @@ class TaskTestCase(TestCase):
 
     def test_task_refresh_data_method_wrong_pk(self):
         logging.disable(logging.WARNING)
-        with self.assertRaises(Ignore):
-            run_model_object_method(
+        run_model_object_method.apply(
+            (
                 self.element._meta.app_label,
                 self.element._meta.model_name,
                 99999,
                 "refresh_data",
             )
+        )
         self.assertEqual(Layer.objects.count(), 0)
 
     def test_task_wrong_method(self):
         logging.disable(logging.ERROR)
-        with self.assertRaises(Ignore):
-            run_model_object_method(
+        run_model_object_method.apply(
+            (
                 self.element._meta.app_label,
                 self.element._meta.model_name,
                 self.element.pk,
                 "bad_method",
             )
+        )
         self.assertEqual(Layer.objects.count(), 0)
 
     @mock.patch("django_geosource.models.Source.objects")
     def test_task_good_method_error(self, mock_source):
         mock_source.get.side_effect = ValueError
         logging.disable(logging.ERROR)
-        with self.assertRaises(Ignore):
-            run_model_object_method(
+        run_model_object_method.apply(
+            (
                 self.element._meta.app_label,
                 self.element._meta.model_name,
                 self.element.pk,
                 "refresh_data",
             )
+        )
         self.assertEqual(Layer.objects.count(), 0)
