+++ test_geosource/celery.py
@@ -4,4 +4,4 @@ app = Celery("django_geosource")
 
 app.config_from_object("django.conf:settings", namespace="CELERY")
 
-app.autodiscover_tasks(["django_geosource.celery"])
+app.autodiscover_tasks()
